Move EmptyStepProvider under test namespace
